+++ cantera18/branches/scons/Cantera/src/thermo/MetalSHEelectrons.cpp
@@ -241,7 +241,7 @@ namespace Cantera {
    * by which the generalized concentration is normalized to produce 
    * the activity. 
    */ 
-  doublereal MetalSHEelectrons::standardConcentration(int k) const {
+  doublereal MetalSHEelectrons::standardConcentration(size_t k) const {
     return 1.0;
   }
   //====================================================================================================================
@@ -249,7 +249,7 @@ namespace Cantera {
    * Returns the natural logarithm of the standard 
    * concentration of the kth species
    */
-  doublereal MetalSHEelectrons::logStandardConc(int k) const {
+  doublereal MetalSHEelectrons::logStandardConc(size_t k) const {
     return 0.0;
   }
   //====================================================================================================================
