+++ cantera18/branches/scons/Cantera/src/thermo/Nasa9PolyMultiTempRegion.h
@@ -121,7 +121,7 @@ namespace Cantera {
     virtual int reportType() const;
 
     //! Returns an integer representing the species index
-    virtual int speciesIndex() const;
+    virtual size_t speciesIndex() const;
  
     //! Update the properties for this species, given a temperature polynomial
     /*!
@@ -202,7 +202,7 @@ namespace Cantera {
      *        coeffs[index+1] = maxTempZone
      *        coeffs[index+2+i] from i =0,9 are the coefficients themselves
      */
-    virtual void reportParameters(int &n, int &type,
+    virtual void reportParameters(size_t &n, int &type,
 				  doublereal &tlow, doublereal &thigh,
 				  doublereal &pref,
 				  doublereal* const coeffs) const;
@@ -222,7 +222,7 @@ namespace Cantera {
     //! standard-state pressure
     doublereal m_Pref;     
     //! species index
-    int m_index;  
+    size_t m_index;
     //! Number of temperature regions
     size_t m_numTempRegions;
